refactor: Replace language-provider with global-translation-provider in multiple components
