+++ components/Providers.tsx
@@ -1,4 +1,3 @@
-import { LanguageProvider } from "@/components/language-provider";
 import { ThemeProvider } from "next-themes";
 import { DiagramExpansionProvider } from "@/lib/hooks/useDiagramExpansion";
 import { AuthProvider } from "@/lib/hooks/use-auth";
@@ -7,16 +6,14 @@ export default function Providers({ children }: { children: React.ReactNode }) {
   return (
     <AuthProvider>
       <DiagramExpansionProvider>
-        <LanguageProvider>
-          <ThemeProvider
-            attribute="class"
-            defaultTheme="system"
-            enableSystem
-            disableTransitionOnChange
-          >
-            {children}
-          </ThemeProvider>
-        </LanguageProvider>
+        <ThemeProvider
+          attribute="class"
+          defaultTheme="system"
+          enableSystem
+          disableTransitionOnChange
+        >
+          {children}
+        </ThemeProvider>
       </DiagramExpansionProvider>
     </AuthProvider>
   );
