+++ components/crowdsource/crowdsource-page-content.tsx
@@ -4,7 +4,7 @@
 
 "use client";
 
-import { useLanguage } from "@/components/language-provider";
+import { useLanguage } from "@/components/global-translation-provider";
 import { CrowdsourceLayout } from "./crowdsource-layout";
 
 /**
