+++ components/how-it-works.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { FeatureStep } from "@/components/feature-step";
-import { useLanguage } from "@/components/language-provider";
+import { useLanguage } from "@/components/global-translation-provider";
 import {
   LinkIcon,
   AudioWaveformIcon,
